remove tim-server module
