+++ tim-group/src/main/java/com/tim/group/TimGroupApplication.java
@@ -1,0 +1,15 @@
+package com.tim.group;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@EnableDiscoveryClient
+@SpringBootApplication
+public class TimGroupApplication {
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(TimGroupApplication.class, args);
+    }
+
+}
