--- tim-server/src/main/java/com/tim/server/common/ChannelManager.java
@@ -1,15 +1,0 @@
-package com.tim.server.common;
-
-import io.netty.channel.Channel;
-import java.util.List;
-
-public interface ChannelManager {
-
-    void addUid2Channel(String uid, Channel channel);
-
-    List<Channel> getChannelByUid(String uid);
-
-    String getUidByChannel(Channel channel);
-
-    void removeChannel(Channel channel);
-}
