+++ tim-single/src/main/java/com/tim/single/common/ChannelManager.java
@@ -1,0 +1,15 @@
+package com.tim.single.common;
+
+import io.netty.channel.Channel;
+import java.util.List;
+
+public interface ChannelManager {
+
+    void addUid2Channel(String uid, Channel channel);
+
+    List<Channel> getChannelByUid(String uid);
+
+    String getUidByChannel(Channel channel);
+
+    void removeChannel(Channel channel);
+}
